apps/train_model/requirements.txt
accelerate
fastapi
huggingface-hub
langchain
langchain-community
langchain-text-splitters
numpy
onnx
onnxruntime
optimum
optimum-onnx
pypdf
python-dotenv
python-multipart
qdrant-client
sentence-transformers
tokenizers
transformers
uvicorn